Created data sampler program and shortened dataset
To run the data sampler place higgsSamplerPython.py into the same folder with HIGGS.csv which is the full dataset file.
The sampler will generate a random seed, use that to generate a random number from one to a million, then use that random number as a seed, then sample 10000 numbers from 1 to 11 million. This sample of 10000 values are the lines of the database which will be taken to form a smaller database.
The program scans through every line of the database, one line at a time, and counts the number of the line currently scanned, if the line number exists in the sample then it writes that line into the new dataset file.